changed db structure to only have users table with role instead of builder/trade table

+++ server/controllers/builder.js
@@ -4,38 +4,39 @@ const router = express.Router()
 
 router.get("/jobs/:id", (req, res) => {
     const sql = `
-        SELECT address, job_id FROM jobs WHERE builder_id = $1
+        SELECT address, job_id FROM jobs WHERE user_id = $1
     `
-    const builder_id = req.params.id
+    const user_id = req.params.id
 
-    db.query(sql, [builder_id]).then((dbResponse) => {
+    db.query(sql, [user_id]).then((dbResponse) => {
         res.json(dbResponse.rows)
     })
 })
 
 router.post("/jobs/:id", (req, res) => {
-    const builder_id = req.params.id
+    const user_id = req.params.id
     const addressObj = req.body
     const address = `${addressObj.suburb}, ${addressObj.lotNumber} ${addressObj.streetName}, ${addressObj.city}, ${addressObj.postcode}`
     console.log(address)
 
     const sql = `
-        INSERT INTO jobs(address, builder_id) VALUES($1, $2)
+        INSERT INTO jobs(address, user_id) VALUES($1, $2)
     `
 
-    db.query(sql, [address, builder_id]).then((dbResponse) => {
+    db.query(sql, [address, user_id]).then((dbResponse) => {
         res.status(200)
         res.json({ success: true })
     })
 })
 
-router.delete("/jobs/:id", (req, res) => {
-    const job_id = req.params.id
+router.delete("/jobs/:job_id/user/:user_id", (req, res) => {
+    const job_id = req.params.job_id
+    const user_id = req.params.user_id
     const sql = `
-        DELETE FROM jobs WHERE job_id = $1
+        DELETE FROM jobs WHERE job_id = $1 AND user_id = $2
     `
 
-    db.query(sql, [job_id]).then((dbResponse) => {
+    db.query(sql, [job_id, user_id]).then((dbResponse) => {
         res.status(200)
         res.json({ success: true })
     })
